PETBlog/blog/requirements.txt
asgiref==3.7.2
build==1.0.3
colorama==0.4.6
crispy-bootstrap5==0.7
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.4
django-crispy-forms==2.0
djangorestframework==3.14.0
filelock==3.12.2
gunicorn==21.2.0
packaging==23.2
Pillow==10.0.1
platformdirs==3.10.0
psycopg2==2.9.7
pyproject_hooks==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.3
whitenoise==6.5.0
